added diversity example

+++ rlflow/utils/space_wrapper.py
@@ -5,7 +5,7 @@ class SpaceWrapper:
     def __init__(self, space):
         if isinstance(space, gym.spaces.Discrete):
             self.shape = ()
-            self.dtype = np.dtype(np.int32)
+            self.dtype = np.dtype(np.int64)
         elif isinstance(space, gym.spaces.Box):
             self.shape = space.shape
             self.dtype = np.dtype(space.dtype)
